provider-project-backend/requirements.txt
Django==2.2.5
PyJWT==1.6.4
mysqlclient==1.4.4
netifaces==0.10.9
Pillow==6.1.0
pkg-resources==0.0.0
PyMySQL==0.9.3
pytz==2019.2
sqlparse==0.3.0
django-cors-headers==3.0.2
djangorestframework==3.9.4
mysql-connector-python==8.0.11
simplejson==3.15.0
SecretStorage==2.3.1
